requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
mysql-connector-python
google-cloud-pubsub
